+++ Models/ApiKeyConstants.cs
@@ -1,0 +1,7 @@
+namespace Review.Api.Models;
+
+public static class ApiKeyConstants
+{
+    public const string ApiKeyAuthorizationScheme = "ApiKey";
+    public const string ApiKeyName = "ApiKey";
+}